requirements.txt
discord==2.3.2
discord.py==2.4.0
easy-pil==0.3.3
python-dotenv==1.0.1